misc: restructured project and added to sections
Splitting the piece.md file into several files makes it easier to
track changes in different parts of the text. The README file,
besides providing a summary of the project, now also serves as a table
of contents listing and linking the several files of the project.

Added mostly to the section about the implementation of looping in
Nodezator, discussing how it can already handle iteration of items,
needing only a small support feature to aid it. Also signaled how
I actually changed my original idea of not providing extra tools
for looping in Nodezator and now plan to do provide such extra tools,
though I have yet to write about this, since I only mentioned it.

I also finished the section on the approach to implement branching
in Nodezator.