added: proposal of nodes emulating for-loops
Changed my stance on not needing to implement nodes to emulate
for-loops in the graph. I now think they are actually helpful in a
lot of situations. As a result, present the idea of the creation of
two new nodes to assist in emulating for-loops in the node graph.

Also created new .svg images to help demonstrate nodes and their usage.